// src/app/shared/models/dsa.ts
export class Dsa {
    profilePicUrl: string;
    currentBadgeId: number
    aadhaarCardNo: string
    businessAddress: string
    businessName: string
    cityId: number
    cityName: string
    commitment: number
    companyName: string
    companyRegNo: string
    companyTypeId: string
    contactNo: string
    createdBy: number
    createdDate: Date
    designationId: null
    fullName: string
    gender: string
    gstNo: string
    haveOffice: boolean
    id: number
    isActive: boolean
    isDelete: boolean
    isDisabled: boolean
    modifiedBy: number
    modifiedDate: Date
    panCardNo: string
    parentPartnerId: number
    permanentCode: string
    professionTypeId: number
    referralCode: string
    temporaryCode: string
    udhyamAadhaarNo: string
    userId: number
    workExperience: number
    totalApplicationReceived: number
    thisMonthAmount: number
    totalAmount: number
    code: string;
    isSelected: boolean = false;
    commission: any;
    partnerAddressId: number
    partnerEducationId: number
    address: string
    pincode: string
    city: string
    districtName: string
    state: string
    profession: string
    roleId: number
    email: string
    addressLine1: string
    addressLine2: string
    roleName: string
    label: string
    resume: string
    documentType: string
    fileName: string
    educationTypeId: string
    instituteName: string
    passingYear: number
    partnerId: number
    documents = []
    badgeName: string
    bankId: number
    accountNo: number
    accountHolderName: string
    ifscCode: string
    partnerBankDetailId: number
    partnerBankDetail: any
    businessAddressLine1: string
    businessAddressLine2: string
    businessPincode: string
    workAddressCityId: number
    workAddressCity: string;
    basicDetail: any;
    partners: any;
    visitingCard:any;
    loans:any;
    employee:any;
    group:any;
    assignTrainings:any;
    walletHistory:any;
    partnerCommission:any
    customers:any;
    orders:any;
    leads:any

    educations:any[]
    
    constructor() { }
}