// src/app/shared/models/users/user.ts
export class Users {
    contactNo: string
    countryCode: string
    createdBy: number
    createdDate: Date
    email: string
    fullName: string
    gender: string
    id: number
    isActive: boolean
    isBlocked: boolean
    isDelete: boolean
    isDisabled: boolean
    isEmailVerified: boolean
    isPasswordSet: boolean
    modifiedBy: number
    modifiedDate: Date
    password: string
    profilePicUrl: string
    roleId: number
    roleName: string
    designation:string;
    permissionGroupId:number
    constructor() { }
}